git-svn-id: file:///home/gecchi/repos@1557 9df8dbd1-7660-4e89-b1e7-fa117e7ea0c7

+++ GgafDx9Core/src/jp/ggaf/dx9core/util/GgafDx9SplineProgram.cpp
@@ -38,9 +38,9 @@ void GgafDx9SplineProgram::behave() {
             return;
         }
 
-        _pActor_target->_X = _sp->_X_compute[SPPointIndex];
-        _pActor_target->_Y = _sp->_Y_compute[SPPointIndex];
-        _pActor_target->_Z = _sp->_Z_compute[SPPointIndex];
+        _pActor_target->_X = (int)(_sp->_X_compute[SPPointIndex]);
+        _pActor_target->_Y = (int)(_sp->_Y_compute[SPPointIndex]);
+        _pActor_target->_Z = (int)(_sp->_Z_compute[SPPointIndex]);
 
         _SPframe++;
